refactor dashboard useEffect to components.
